% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{allele.qc}
\alias{allele.qc}
\title{Identify incorrectly coded and reversed alleles}
\usage{
allele.qc(a1, a2, ref1, ref2)
}
\arguments{
\item{a1, a2}{vectors of effect / reference allele from first source}

\item{ref1, ref2}{vectors of effect / reference alleles from second source
(implied: the reference, but not necessary)}
}
\value{
A list with two logical vectors, each of the same length as the
put vectors: \code{keep} indicating SNPs that have valid coding and are
not strand-ambiguous, and \code{flip} indicating SNPs with reversed
effect / reference definition between sources.
}
\description{
When comparing effect / reference alleles for vectors of SNPs from two sources,
identify a. SNPs that are either not ACGT or strand-ambiguous, b. where effect
/ reference allele have been reversed between sources.
}
