% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtwas.R
\name{condTWAS_snp}
\alias{condTWAS_snp}
\title{Condition per-SNP GWAS statistics on per-gene TWAS statistics}
\usage{
condTWAS_snp(
  summarystats,
  genelist,
  refdata,
  pred_ge,
  pred_ge_corr,
  cons_loci,
  lambda = 0,
  opts = opts_rtwas$get()
)
}
\arguments{
\item{opts}{List of options used for extracting default values for unspecified
arguments}
}
\description{
Given GWAS summary statistics for a chromosome and a corresponding set of
TWAS test statistics, this function will calculate adjusted versions of the
summary statistics, conditional on the observed TWAS statistics.
}
