% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtwas.R
\name{eqtl_to_refdat}
\alias{eqtl_to_refdat}
\title{Map EQTL gene data to reference data}
\usage{
eqtl_to_refdat(genelist, refdata)
}
\arguments{
\item{genelist}{A data frame with information on the genes of interest,
including the path to the data file containing the expression
weights; typically output from \code{read_genelist}}

\item{refdata}{Genetic reference data, as a list with PLINK-style components;
typically output from \code{read_refdata}}
}
\value{
A list with two components:\preformatted{* `refsnps_keep`: a logical vector that indicates for each SNP in the
                  reference data whether it has a weight associated in
                  the EQTL data.
* `pred_expression`: a numerical matrix with the predicted gene expression
                     component for each gene (as columns) for all subjects
                     in the reference data (as rows).
}
}
\description{
Given a list of genes and the weights for their EQTL-SNPs, this function maps
the EQTL-SNPs to SNP-level reference data, and predicts all genetic expression
components in the expression data.
}
\details{
This is a bit of a mixed function, in that the two result components
are only loosely related, and would logically be calculated in separate
functions. However, calculation of both requires iterating over the per-gene
expression weights, which are stored as files, so that would not be very
efficient.
}
\seealso{
\code{\link[=read_genelist]{read_genelist()}}, \code{\link[=read_refdata]{read_refdata()}}
}
