% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expand_special}
\alias{expand_special}
\title{Expand file names}
\usage{
expand_special(template, chr)
}
\arguments{
\item{template}{character vector with template or file names}

\item{chr}{a numerical vector of chromosome numbers (range 1-22)}
}
\value{
A character vector of file names, expanded if the input \code{template}
contained the special \verb{!!!!}, otherwise the original input
}
\description{
Replace special string \verb{!!!!} with the number of the chromosomes to be
processed in file name templates
}
\examples{
expand_special("sumstats_chr!!!!", 1:22)
expand_special("sumstats_wholeGenome", 1:22)
expand_special(c("sumstats_chr1", "sumstats_chr2"), 1:2)
}
