% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_loci}
\alias{find_loci}
\title{Identify contiguous loci around EQTL-SNPs in the genetic reference data}
\usage{
find_loci(refdata, snp_ndx, opts = opts_rtwas$get())
}
\arguments{
\item{refdata}{Genetic reference data, as a list with PLINK-style components;
typically output from \code{read_refdata}}

\item{snp_ndx}{logical vector indicating for each SNP in the reference data
whether it carries an EQTL weight (and is therefore included
in the locus-construction); typically output from \code{eqtl_to_refdat}.}
}
\value{
A data frame with columns \code{starts} and \code{ends} indicating the start and
end base pair position of each locus
}
\description{
For the conditional TWAS analysis, both the SNPs and the conditioning
(gene-level) TWAS-statistics are clumped into contiguous loci.
}
\seealso{
\code{\link[=read_refdata]{read_refdata()}}, \code{\link[=eqtl_to_refdat]{eqtl_to_refdat()}}
}
