% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\docType{data}
\name{opts_rtwas}
\alias{opts_rtwas}
\title{Options for the current RTWAS analysis}
\format{
An object of class \code{list} of length 6.
}
\usage{
opts_rtwas
}
\description{
The FUSION code base on which RTWAS is built uses an option list to control
the execution of the component scripts: for specifying input data files,
output directories, reference data, type of analysis, window sizes for genetic
loci, plot legends etc. RTWAS initializes and maintains a compatible
list of options, both defaults and required inputs, but uses only a subset
of options itself (e.g. names of input files), while type of analysis or plot
are determined by calling the corresponding function rather than setting an
option.
}
\details{
The management of the option list is built on the \code{new_defaults}-mechanism
implemented in package \code{knitr}: the list of default options is an R6 object
with main methods \code{get} and \code{set} to either read or modify the default values.
The basic list of options \code{opts_rtwas} is a global object that lives in the
package namespace, though local instances (for multiple concurrent analyses)
can be created via calls to \code{knitr:::new_defaults} and the \verb{$restore}-method.

Normally, we would start an analysis by setting the data- and output options
in a call to \code{opts_rtwas$set}, possibly followed by a second call to modify
default options for the desired function calls.

FIXME: a list of options
}
\examples{
opts_rtwas$get("chr")
opts_rtwas$set(chr = 17)
opts_rtwas$get("chr")
## Complete list
opts_rtwas$get()
}
\keyword{datasets}
