% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtwas.R
\name{postTWAS_conditional_chr}
\alias{postTWAS_conditional_chr}
\title{Run a conditional TWAS analysis for a single chromosome}
\usage{
postTWAS_conditional_chr(
  chr,
  fn_genelist,
  fn_sumstats,
  fn_output,
  lambda = 0,
  opts = opts_rtwas$get()
)
}
\arguments{
\item{chr}{Number of chromosome to analyse (see Details).}

\item{fn_genelist}{Name of file listing genes with TWAS statistics for
conditional analysis}

\item{fn_sumstats}{Name of file listing the GWAS summary statistics to be
conditioned}

\item{fn_output}{Name of file for output generated by this function; note that
this is a stub, potentially shared by multiple file names
differing by an extension of the stub}

\item{lambda}{Ridge parameter for calculating the conditional GWAS statistics
(experimental); default zero, i.e. same behavior as FUSION}

\item{opts}{List of options used for extracting default values for unspecified
arguments}
}
\value{
The same list of results as \code{condTWAS_snp}(the worker function).
}
\description{
This is a wrapper function that combines all necessary function calls to
perform a post-TWAS conditional analysis (at the SNP/locus level).
}
\details{
If file names (options \verb{fn-}) arenot specified, the function falls
back on the corresponding elements of the global option argument, namely
\code{ìnput}, \code{sumstats} and \code{out} (see \code{opts_rtwas}); if these are also not set,
the function will fail with an error.

Note that strictly speaking, the \code{chr}-argument is redundant, as this is
implicitly determined by specifying chromosome-specific input files; this is
more of a fallback for situations where the either the gene list- or summary
statistics files are not single-chromosome.
}
\seealso{
\code{\link[=condTWAS_snp]{condTWAS_snp()}} \link{opts_rtwas}
}
