% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtwas.R
\name{postTWAS_conditional_gw}
\alias{postTWAS_conditional_gw}
\title{Run a genome-wide conditional TWAS analysis f}
\usage{
postTWAS_conditional_gw(
  fn_genelist,
  fn_sumstats,
  fn_output,
  chr = 1:22,
  lambda = 0,
  opts = opts_rtwas$get()
)
}
\arguments{
\item{fn_genelist}{Name of file listing genes with TWAS statistics for
conditional analysis; see Details}

\item{fn_sumstats}{Name of file listing the GWAS summary statistics to be
conditioned; see Details}

\item{fn_output}{Name of file for output generated by this function; note that
this is a stub shared by multiple output file names
differing by an extension of the stub}

\item{chr}{Numbers of chromosomes to analyse, by default \code{1:22}}

\item{lambda}{Ridge parameter for calculating the conditional GWAS statistics
(experimental); default zero, i.e. same behavior as FUSION}

\item{opts}{List of options used for extracting default values for unspecified
arguments}
}
\value{
A list of per-chromosome post-conditional TWAS results, as returned
by \code{postTWAS_conditional_chr} (which are essentially the same results as
returned by \code{condTWAS_snp}).
}
\description{
This is a wrapper function that combines all necessary function calls to
perform a post-TWAS conditional analysis (at the SNP/locus level) across
a set of chromosomes (by default, all chromosomes 1-22, i.e. genome-wide).
}
\details{
File names specifying where to read the genes with a TWAS statistic
of interest (\code{fn_genelist}), the corresponding GWAS summary statistics
(\code{fn_sumstats}) and a stub for output file names (\code{fn_output}) need to
be provided via the named arguments, whereas the location of the reference
data need to be specified via \code{opts_rtwas}.

There is some flexbility in how TWAS- and GWAS statistics can be
specified, however:
\enumerate{
\item As character vectors: this puts the responsibility for naming files
completely on the user - they need to provide matching vectors of
fully specified \code{genelist}- and \code{sumstats} file names that correspond
to the list of chromosomes to be analysed.
\item As character expressions of length 1: these are interpreted as file name
\emph{templates}: the list of file names will be constructed by replacing
the special string \verb{!!!!} with the numbers of the chromosomes specified
in argument \code{chr}.
}

Note that the same special string can (and generally \emph{should}) be used for
the file stub for output file names. All copies of \verb{!!!!} will be replaced,
so this can be used to e.g. have both chromosome-specific file names and
chromosome-specific subfolders.
}
\seealso{
\code{\link[=postTWAS_conditional_chr]{postTWAS_conditional_chr()}} \code{\link[=condTWAS_snp]{condTWAS_snp()}} \link{opts_rtwas}
}
