% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtwas.R
\name{predict_expression}
\alias{predict_expression}
\title{Predict expression on genetic reference data}
\usage{
predict_expression(genelist, refdata)
}
\arguments{
\item{genelist}{A data frame with information on the genes of interest,
including the path to the data file containing the expression
weights; typically output from \code{read_genelist}}

\item{refdata}{Genetic reference data, as a list with PLINK-style components;
typically output from \code{read_refdata}}
}
\value{
A matrix with subjects in the reference data as rows and genes as
columns, containing for each subject / gene the predicted expression level
(scaled across ubjects).
}
\description{
Given a list of genes and the expression weights for their eQTLs,
calculate the matrix of predicted gene expression values on a
genetic reference data set.
}
\seealso{
\code{\link[=read_genelist]{read_genelist()}}, \code{\link[=read_refdata]{read_refdata()}}
}
