% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_genelist}
\alias{read_genelist}
\title{Read list of genes with TWAS z-statistics}
\usage{
read_genelist(file, chr, minp, opts = opts_rtwas$get())
}
\arguments{
\item{file}{Name of the file with gene names}

\item{chr}{Number of chromosome to use for filtering (\code{NA} for no filtering)}

\item{minp}{Minimal significance level required to be included (\code{NA} for no filtering)}

\item{opts}{List of options used for extracting default values for unspecified
arguments}
}
\value{
A data frame with rows as genes and columns containing gene information
and TWAS results FIXME: more details
}
\description{
Read the output file with the gene-level TWAS statistics as generated by
script FUSION.assoc_test.R for further testing / conditional TWAS.
}
