% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_refdata}
\alias{read_refdata}
\title{Read genetic reference data}
\usage{
read_refdata(file, chr, scale = TRUE, opts = opts_rtwas$get())
}
\arguments{
\item{file}{Stub of the file names containing the reference data, see Details.}

\item{chr}{Number of chromosome to read in}

\item{scale}{Logical flag indicating whether to scale the allele counts
(default \code{TRUE})}

\item{opts}{List of options used for extracting default values for unspecified
arguments}
}
\value{
FIXME
}
\description{
Read the genetic reference data for establishing LD estimates etc.
}
\details{
The reference data is assumed to be in PLINK format (.bed/.bim/.fam),
with file names as \verb{<stub>.<chr number>.<ext>}. The standard data used by
FUSION is the European subset of the 1000 Genomes data.
}
