% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_sumstats}
\alias{read_sumstats}
\title{Read GWAS summary statistics}
\usage{
read_sumstats(refdata, opts = opts_rtwas$get())
}
\arguments{
\item{refdata}{Genetic reference data, as a list with PLINK-style components;
typically output from \code{read_refdata}}

\item{opts}{List of options used for extracting default values for unspecified
arguments}
}
\value{
FIXME
}
\description{
Read the summary statistics for the phenotype of interest
}
\details{
data format FIXME
}
